# api.yml
openapi: 3.0.3
info:
  title: Spring API
  version: 1.0.0
paths:
...
  /trips:
    get:
      parameters: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items: &ref_14
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    moveRequestId:
                      type: integer
                      format: int64
                    moveOfferId:
                      type: integer
                      format: int64
                    status:
                      type: string
    post:
      parameters: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *ref_14
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                moveRequestId:
                  type: integer
                  format: int64
                moveOfferId:
                  type: integer
                  format: int64
                status:
                  type: string
            example:
              moveRequestId: 0
              moveOfferId: 0
              status: ''
  /trips/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *ref_14
  /auth/login:
    post:
      parameters: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &ref_15
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
                  userId:
                    type: integer
                    format: int64
                  name:
                    type: string
                  email:
                    type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
            example:
              email: ''
              password: ''
  /auth/register:
    post:
      parameters: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *ref_15
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                dob:
                  type: string
                  format: date
                role:
                  type: string
            example:
              username: ''
              email: ''
              password: ''
              firstName: ''
              lastName: ''
              dob: '2026-03-08'
              role: ''
...
  /auth/check:
    get:
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /auth/me:
    get:
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  username:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  passwordHash:
                    type: string
                  dob:
                    type: string
                    format: date
                  rating:
                    type: number
                    format: float
                  role:
                    type: string
                  addressId:
                    type: integer
                    format: int64